Allow the GHCi Linker to resolve related dependencies when loading DLLs
Summary:
GHCi does not correctly tell the Windows Loader how to handle dependencies to dll's that are not on the standard windows load path:

1. The directory from which the application loaded.
2. The current directory.
3. The system directory. Use the GetSystemDirectory function to get the path of this directory.
4. The 16-bit system directory. There is no function that obtains the path of this directory, but it is searched.
5. The Windows directory. Use the GetWindowsDirectory function to get the path of this directory.
6. The directories that are listed in the PATH environment variable. Note that this does not include the per-application path specified by the App Paths registry key. The App Paths key is not used when computing the DLL search path.

So what this means is given two DLLs `A` and `B` and `B` depending on `A`. If we put both DLLs into a new folder bin and then call GHC with:

`ghc -L$(PWD)/bin -lB`

the loading will fail as the Windows loader will try to load the dependency of B and fail since it cannot find A.

*IMPORTANT* this patch drops XP Support.
Also the APIs being used were natively added to Windows 8+ and backported to Windows 7 and Vista via a mandatory security patch (in 2011).
This means that there is a chance that KB2533623 has not been installed on certain machines. For those machines I display a warning
and temporarily expand the `PATH` to allow it to load.

Test Plan:
./validate

Added two new test T10955 and T10955dyn

Reviewers: austin, thomie, bgamari, hvr, goldfire

Subscribers: erikd, #ghc_windows_task_force

Differential Revision: https://phabricator.haskell.org/D1340

GHC Trac Issues: #10955