Exec: workaround for execv behaviour on Windows.
Summary:
In a land far far away, a project called Cygwin was born.
Cygwin used newlib as it's standard C library implementation.

But Cygwin wanted to emulate POSIX systems as closely as possible.
So it implemented `execv` using the Windows function `spawnve`.

Specifically

```
spawnve (_P_OVERLAY, path, argv, cur_environ ())
```

`_P_OVERLAY` is crucial, as it makes the function behave *sort of*
like execv on linux. the child process replaces the original process.

With one major difference because of the difference in process models
on Windows: the original process signals the caller that it's done.

this is why the file is still locked. because it's still running,
control was returned because the parent process was destroyed,
but the child is still running.

I think it's just pure dumb luck, that the older runtimes are slow enough
to give the process time to terminate before we tried deleting the file.
Which explains why you do have sporadic failures even on older runtimes
like 2.5.0, of a test or two (like T7307).

So this patch fixes a couple of things. I leverage the existing `timeout.exe`
to implement a workaround for this issue.

a) The old timeout used to start the process then assign it to the job.
   This is slightly faulty since child processes are only assigned to a
   job is their parent were assigned at the time they started. So this was
   a race condition. I now create the process suspended, assign it to the job
   and then resume it. Which means all child processes are not running under
   the same job.

b) First things, Is to prevent dangling child processes. I mark the job with
   `JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE` so when the last process in the job
   is done, it insures all processes under the job are killed.

c) Secondly, I change the way we wait for results. Instead of waiting for the
   parent process to terminate, I wait for the job itself to terminate.

   There's a slight subtlety there, we can't wait on the job itself. Instead
   we have to create an I/O Completion port and wait for signals on it.
   See https://blogs.msdn.microsoft.com/oldnewthing/20130405-00/?p=4743

This fixes the issues on all runtimes for me and makes T7307 pass consistenly.

Test Plan: ./validate

Reviewers: RyanGlScott, bgamari, austin, erikd

Subscribers: thomie, #ghc_windows_task_force

Differential Revision: https://phabricator.haskell.org/D2684

GHC Trac Issues: #12725, #12554, #12661, #12004